requirements.txt
llama-cpp-python==0.2.20
PyPDF2==3.0.1
python-docx==1.1.0
python-dotenv==1.0.0
huggingface-hub==0.19.4
transformers==4.35.2
torch==2.1.0
accelerate==0.24.1
peft==0.6.0
bitsandbytes==0.41.3
sentencepiece==0.1.99
protobuf==3.20.3
datasets==2.14.6
scipy==1.11.4
safetensors==0.4.0